# Have calculate_wait stop just shy of 1sec offset to optimize distribution