The step size between calls is 1s and the random offset is called to avoid waves, so have it stop just shy of 1s. This addresses the random failing test in test_http caused when the random offset was 1s.

**Related Issue(s):**

Closes #723


**Proposed Changes:**

For inclusion in changelog (if applicable):

1. 

Not intended for changelog:

1. Have calculate_wait stop just shy of 1sec offset

**Diff of User Interface**

Old behavior:

running the test 5000 times resulted in 37 failed tests


New behavior:

running the test 5000 times resulted in 0 failed tests


**PR Checklist:**

- [x] This PR is as small and focused as possible
- [x] If this PR includes proposed changes for inclusion in the changelog, the title of this PR summarizes those changes and is ready for inclusion in the Changelog.
- [x] I have updated docstrings for function changes and docs in the 'docs' folder for user interface / behavior changes
- [x] This PR does not break any examples or I have updated them

**(Optional) @mentions for Notifications:**