@jreiberkyle 2023-01-20
LOL it turns out the test was actually telling us something (that we shouldn't be going up to 1000s for optimal distribution) instead of just failing. Good test!